Consolidated test code into base class
